[RAPTOR-8184] A merge base commit should be searched against a remote branch
